[FLINK-34078][state]Move InternalKeyContext classes from o.a.f.runtime.state.heap to o.a.f.runtime.state package
