+++ flink-state-backends/flink-statebackend-changelog/src/main/java/org/apache/flink/state/changelog/restore/KvStateChangeApplier.java
@@ -18,8 +18,8 @@
 package org.apache.flink.state.changelog.restore;
 
 import org.apache.flink.core.memory.DataInputView;
+import org.apache.flink.runtime.state.InternalKeyContext;
 import org.apache.flink.runtime.state.KeyGroupRangeAssignment;
-import org.apache.flink.runtime.state.heap.InternalKeyContext;
 import org.apache.flink.runtime.state.internal.InternalKvState;
 import org.apache.flink.runtime.state.internal.InternalMergingState;
 import org.apache.flink.state.changelog.StateChangeOperation;
